fix(api/product/[id]): fix order controllers y route
