+++ app/api/products/[id]/controllers.js
@@ -3,7 +3,7 @@ import prisma from "../../db/client";
 const getProduct = async (id) => {
   const searchedUser = await prisma.Product.findFirst({
     where: {
-      id: id,
+      id: +id,
     },
     include: {
       reviews: true,
@@ -24,19 +24,38 @@ const getProduct = async (id) => {
   return searchedUser;
 };
 
+const updateProduct = async (id, name, description, price, stock, image) => {
+  if (!id && !name && !description && !price && !stock && !image)
+    throw new Error("Missing argument");
+
+  const product = await prisma.product.update({
+    where: {
+      id: +id,
+    },
+    data: {
+      name: name,
+      description: description,
+      price: price,
+      stock: stock,
+      image: image,
+    },
+  });
+  return product;
+};
+
 const deleteProduct = async (id) => {
   const searchedUser = await prisma.Product.findFirst({
     where: {
-      id: id,
+      id: +id,
     },
   });
   if (!searchedUser) throw new Error("Product doesn't exist");
   const searchIdProd = await prisma.Product.delete({
     where: {
-      id: id,
+      id: +id,
     },
   });
   return searchIdProd;
 };
 
-export { getProduct, deleteProduct };
+export { getProduct, deleteProduct, updateProduct };
