+++ app/api/products/[id]/route.js
@@ -1,8 +1,8 @@
 import { NextResponse } from "next/server";
-import { getProduct, deleteProduct } from "./controllers";
+import { getProduct, deleteProduct, updateProduct } from "./controllers";
 
 export async function GET(request, { params }) {
-  const id = +params.id;
+  const { id } = params;
 
   try {
     const response = await getProduct(id);
@@ -12,12 +12,23 @@ export async function GET(request, { params }) {
   }
 }
 
-export async function DELETE(request, { params }) {
-  const id = +params.id;
+export async function PUT(request, { params }) {
+  const { id } = params;
+  const { name, description, price, stock, image } = await request.json();
 
   try {
-    const response = await deleteProduct(id);
+    const response = await updateProduct(id, name, description, price, stock, image);
     return NextResponse.json(response, { status: 200 });
+  } catch (error) {
+    return NextResponse.json({ error: error.message }, { status: 404 });
+  }
+}
+
+export async function DELETE(request, { params }) {
+  const { id } = params;
+  try {
+    await deleteProduct(id);
+    return NextResponse.json({ message: `Product with id: ${id} deleted` }, { status: 200 });
   } catch (error) {
     return NextResponse.json({ error: error.message }, { status: 404 });
   }
